Use imaps for clang_complete's snippets
Using UltiSnips with clang_complete fails sometimes (the selection
expands beyond the function signatures, for example), and the snippets
solution provided by clang_complete requires going to normal mode to
jump to the next placeholder.

Now the snippets are using the 'imaps' plugin provided by LaTeXSuite,
which implements a placeholders system that is similar to the
clang_complete's snippets solution, but <C-j> can be used to jump to the
next placeholder, even in insert mode.