Fix custom imaps clang_complete's snippets
Don't search for the placeholder if there is no need to do it (example:
a function without parameters) and keep the cursor position at the end
of the snippet.

Whitout this fix, the cursor jumps to the beginning of the line (to
search for the first placeholder, that doesn't exist).