Keep window position when switching buffers
By default, switching buffers keeps the cursor position in the same
line, but the line itself may be repositioned in the middle of the
screen.

Add the autocommands to restore the line to its original position
because that is where I expect the line to be, not in a shifted
position.

[1]: http://vim.wikia.com/wiki/Avoid_scrolling_when_switch_buffers
[2]: http://stackoverflow.com/questions/4251533/vim-keep-window-position-when-switching-buffers