Use default formatoptions
Use the Vim distribution values for each filetype, but set it for those
filetypes that doesn't change the Vim default value (e.g. python).

Note: formatting the paragraph using the indentation of the second line
is now removed of the options. I need to figure out in which cases I was
using it to restore it for every filetype individually.