Update command to switch to alternate buffer
<C-^> appears to use :edit, which in turns reevaluates existing
modelines. This causes, for example, that folding state is not
remembered when switching back to a buffer [1].

Replace <C-^> with :buf#, which also switches to the alternate buffer,
but it doesn't load the modeline. (In the folding use-case, the state
is preserved. An open fold will not be closed even if the modeline is
set to close all folds.)

[1]: https://github.com/kien/ctrlp.vim/issues/97