Fix unmapping 'ds' keymap on Buffergator
Commit db0b8ee8a56f ("Improve key mappings on buffer navigation plugin")
attempted to unmap 'ds' set by Surround plugin to avoid conflicts with
the 'd' map used by Buffergator. But the keymap was unmapped globally
instead of just in the buffer, and 'ds' Surround action could not be
used anymore after opening Buffergator.

Since a global map cannot be unmapped for a buffer, ensure the 'ds' map
is restored after leaving the Buffergator plugin.