Trying to remove netrw buffers from buffers list
Add experimental workaround to remove netrw buffers from the buffer
list.

This happens when using vinegar to navigate the directory of the current
file: the netrw buffers may stick around unclosed [1], or the first
netrw buffer will be part of the list of buffers [2]. It is annoying to
navigate buffers with buffergator or ctrlp when there are several netrw
buffers listed along with the open files.

Try setting the bufhidden option to delete the netrw buffers, and use a
custom function to delete that first netrw buffer that is not removed
just by setting bufhidden.

[1] https://github.com/tpope/vim-vinegar/issues/13
[2] https://github.com/tpope/vim-vinegar/issues/74