Increase maximum column for syntax in Git buffers
The option `synmaxcol` is set to 128 by the main `vimrc`, and when a
long line exceeds that limit it stops being highlighted, and the next
lines aren't highlighted properly any more. In other words, after a long
line highlighting is broken and off-sync.

This may happen a lot with verbose Git commits if the patch has a long
diff line (like a "diff --git /path/orig /path/changed" line with long
paths).

So, increase the column to 256 to prevent highlighting issues with
patches inside Git buffers. 256 should be enough for most cases (and it
is greater than full width of my maximized terminal at 200 chars).

Should the value in vimrc be changed instead? 128 seems a bit too
extreme.