Remove nocompatible from gvimrc
It should not be set there anyway, and it was creating issues when
starting GVim (specifically, a custom location to the viminfo file was
removed because the option was reset).

Actually, it shouldn't even be necessary to set this in the vimrc at
all, because it is set automatically when a user vimrc file is
detected [0], but if it is not present there then the startuptime.vim
plugin will fail (it uses `-u vimrc`, and when `-u` is passed then
`nocompatible` is not automatically set).

So, to avoid issues, keep the `nocompatible` line in vimrc, even when
most of the time is a noop.

[0]: http://vimdoc.sourceforge.net/htmldoc/starting.html#compatible-default