Trying to ensure command completion is case sensitive
Activating case insensitive search has the secondary effect of making
command completion also case insensitive. But if I type ":n<TAB>" I only
want to see Vim lowercase commands as the completion alternatives and
not all the plugin-defined commands that start with "N".

This commit is an attempt to solve this (after many years of being
annoyed by the completion) by changing to case sensitive search only for
command line.

So far so good.