Add install script option to just update the plugins
Useful to run after updating the repository to the latest version.
A snapshot is created to store the state of the plugins before updating.

The skip option was replaced with this new update option. The skip
option was confusing, and not really useful. Skip installing plugins but
do download spell files? No need to skip the install in that case, it
would be a no-op most of the time.