Fix making command completion case sensitive
Fix first attempt in 1434b43 (Trying to ensure command completion is
case sensitive, 2019-02-21) by ensuring that 'noignorecase' is set only
when ':' is used. Otherwise all commands that open the command line
would be configured by the autocommand. This meant that '/' and '?'
would start with case sensitive search (and then 'n'/'N' would go back
to case insensitive), which is not the expected behaviour.