Prefer compile database for C++ options
When editing a C++ source file that is not in a custom project ,i.e. no
.project.vim file, set the C++ default options (C++14 standard and all
warnings) only if a compile_commands.json database cannot be
found in the default locations.

This is to prevent the global options to interfere with the options from
the database, because the global options are appended to the full
compilation command.

If there is a custom .project.vim file, then do not modify the global
options. It is expected that the local project file may have set them
already (or the proper location for the compilation database).