Disable airline searchcount extension
It crashes every time a parenthesis is opened when typing a query in the
search prompt (it doesn't even give time to close the parenthesis):

    Error detected while processing function airline#extensions#searchcount#status:
    E54: Unmatched \(

It seems triggered by having the ruler, incsearch and hlsearch options
set.

In any case, I find the extension annoying and it doesn't integrate well
with the theme.