Do not hardcode ~/.vim location
Neovim uses a different configuration directory (which follows XDG),
and mixing directories may result in a weird runtimepath. Also this
could be extended to actually support having the Vim configuration in a
different directory than ~/.vim (a use case that I needed when sharing
an Amazon box with other NASA scientists).

Unfortunately this could not be used:

    let $VIM_CONFIG_HOME = expand('<sfile>:p:h')

since the Vim configuration may be managed by a dotfiles manager and
then symlinked into ~/.vim ($XDG_CONFIG_HOME/nvim). That results in
'<sfile>' getting resolved to its realpath instead of ~/.vim/vimrc
($XDG_CONFIG_HOME/nvim/init.vim), which is not what I want because the
runtimepath may get mixed up again when it is edited to support the
private directory.

The install script still expect to be in ~/.vim though.