Add support to disable loading private configuration
If the environment variable VIM_CONFIG_LOAD_PRIVATE_RC is set to 0,
the private configuration will not be loaded. Useful to debug the default
configuration when there are multiple private runtime files, which makes
it hard to temporary disable them.

Note that `bundle_*.vim` and `vimrc_*` files are not considered
"private" though, because the idea is to support "fork configuration"
that may extend this configuration by adding `bundle_company.vim` or
`vimrc_fork` for example, so those files should not be treated as
private configuration. For the actual private files of the local
machine, like `vimrc_local`, it is recommended to add a check at the
beginning:

    if $VIM_CONFIG_LOAD_PRIVATE_RC == '0'
        finish
    endif

that will exit the file without sourcing the rest of the content.