Increase maximum column for syntax highlighting
Increase synmaxcol to 256 for all file types. It was originally set to 128
for performance reasons by 870d108 (Set max syntax columns to prevent lag
on long lines, 2012-04-30), but I have found too many files where lines go
longer than 128 and then the syntax highlighting breaks for next lines.
Just try setting it to 256, which should cover almost all those cases that
break when it is set to 128; and it should be performant enough anyway,
compared to the default of 3000.

This is a follow up to 402cf0d (Increase maximum column for syntax in Git
buffers, 2018-11-22) and 06033da (Increase maximum column for syntax in
Markdown buffers, 2019-02-21), that increased the value just for Git and
Markdown files.