Fix custom j/k remaps
The j/k keys are remapped to gj/gk to allow easy cursor navigation for
long lines, but if something like 5j is pressed, then the cursor would
jump 5 displayed lines instead of 5 actual lines.

Fix the mapping to ensure that the proper number of lines are jumped
when using a count.