Fix UltiSnipsEdit command (again)
`:UltiSnipsEdit` fails because it cannot find a `.vim` directory. This
happens because UltiSnips now uses `glob` to add snippets directories
from the runtime path, which returns empty for non-existing directories.
Since I don't use `~/.vim/UltiSnips`, it doesn't exist and the command
now fails, and the workaround introduced by cd0c5a2 (Override
:UltiSnipsEdit to prefer private snippets, 2020-08-11) is not enough to
prefer `~/.vim/private/UltiSnips` as the snippets directories for
`:UltiSnipsEdit`.

But the workaround was needed because UltiSnips removed the variable to
set a prefered directory for `:UltiSnipsEdit`. Fortunately, they
recently re-added the variable (although with a different, very long
name), so I can revert the workaround and restore the behavior of
`:UltiSnipsEdit` the way I want it, and everythin works again.