Relocate CoC user configuration
Checking VIM_CONFIG_USE_COC doesn't work when the variable is set but
Vim is started with the `-u {vimrc}` option, and CoC is not added as
plugin by the given file. The check will pass and `coc#config` will be
called, but not found, resulting in an error.

Cannot use a custom flag in `bundle.vim` to mark that CoC is added as
plugin and then check if that flag is set, because the check will pass
even when CoC is not installed yet with `:PlugInstall`, resulting in an
error again.

The best option is to check if CoC was sourced with the variable
`g:did_coc_loaded`. But then cannot use the location `plugin/coc.vim`
because CoC is not sourced yet when that file is sourced.

So, use `g:did_coc_loaded` to check, and move the configuration to
`after/plugin/coc.vim`, which is sourced after plugins, to ensure that
the variable exists when CoC is used.