Try to fix the airline section-z symbols
The section-z is a mess since a couple of PRs changed it to more Unicode
symbols, and manipulated the spacing, for the worse. Now it looks like
this:

     5% ㏑:16/284☰℅:1

Try to fix it to make it look more pleasant, similar to the original
definition before the PRs:

     5% ☰ 16/284:1

Unfortunately this still leaves the status line for `:terminal` with an
extra space, because it is needed in the `linenr` symbol to add a
separation after the `%p`, but this symbol is also used by the term
extension to show the number of lines in `:terminal`.