Create persistent state directories if missing
Relying only on the install script to create the directories is not
robust. The state directory may have been removed for any reason, and
then the persistent files will be created in the current working
directory, because the expected directories are missing.

Prefer checking and creating the directories on initialization.