Remove override directories from setup options
Follow up to 66a700c (Put setup options into a global dictionary,
2022-11-16). Remove the `use_<dir>` variables from the setup dictionary
that can be overriden by the user, since it just complicates things.

Prefer them to be inmutable, depending only if the directories exist or
not. Rename them to `has_<dir>`, since `use_<dir>` can be understood as
configurable.