Force black background colorscheme
Force the default darkglass colorscheme to use a black background when
Vim detects the terminal background as light. Otherwise for terminals
with white background the colors would not look right.

Vim may detect background as light even when the terminal has a dark
background. The correct autodetection by querying the terminal with t_RV
seems to take a bit and finishes after the colorscheme is loaded.
In that case the transparent background will be replaced now by a solid
black background.

The environment variable COLORFGBG can be used to ensure Vim detects a
dark background.