Improve completion trigger map
When pressing CTRL+Space, the actual key code depends on the terminal
and Vim version:

- Old terminals treat CTRL+Space as <C-@>
- GVim and Neovim treat CTRL+Space as actual <C-Space>
- New terminals that support new protocols can also treat CTRL+Space as
  <C-Space> now instead of <C-@> if Vim uses the modern protocol [0]

Currently the first two cases are supported by just using <C-@> for Vim
on a terminal and <C-space> in GVim/Neovim, as the {lhs} of the
completion trigger map. But it's hard to know if the third case is
supported by current terminal instead of the first case [0], so I don't
see a way to add a new condition to detect if <C-@> or <C-Space> should
used as the {lhs} for Vim on terminal.

In order to support all cases, simply map both to trigger completion.
Whichever is supported will be used then.

This also fixes the visual map to expand snippets, which was not working
when using CoC as completion engine.

[0]: Vim 9.1.0969 made Ghostty to see <C-Space> now instead of C-@